refactor: data polling code
