+++ src/components/LanguageDropDown/style.scss
@@ -24,7 +24,8 @@
           right: -15px;
           display: block;
           background: #fff;
-          width: 120px;
+          min-width: 125px;
+          width: max-content;
           padding-top: 0px;
           z-index: 1;
           border-radius: 5px;
@@ -78,7 +79,11 @@
             &:hover {
               color: #146c78;
             }
-    
+          }
+
+          .active {
+            color: #146c78;
+            font-weight: bold;
           }
         }
       }
