+++ src/routes/stay/+page.ts
@@ -3,17 +3,7 @@ import { stays } from "$lib/store/stay.store";
 import { tick } from "svelte/internal";
 import { get } from "svelte/store";
 /** @type {import('./$types').PageLoad} */
-export async function load({fetch} : any ) {
-  await tick();
-  getStays({
-    fetchFunction: fetch,
-    // sample payload
-    stayPayload: {
-      campId: "fafafafa-b2a8-4885-bace-875199e719aa",
-      endDate: 4083077146160,
-      startDate: 0,
-    },
-  });
+export async function load() {
   return {
     stays: get(stays),
   };
