+++ src/routes/user/+page.ts
@@ -1,13 +1,7 @@
-import { getUsers } from "$lib/api/user";
 import { users } from "$lib/store/user.store";
-import { tick } from "svelte/internal";
 import { get } from "svelte/store";
 /** @type {import('./$types').PageLoad} */
-export async function load({ fetch } : any) {
-  await tick();
-  getUsers({
-    fetchFunction: fetch,
-    });
+export async function load() {
   return {
     users: get(users),
   };
